Update login, language-switcher components and add not-found page
Implemented significant user interface updates and additions. These include the rework of the login and language-switcher components, and the introduction of a not-found page. Integrated 'ngx-toastr' and modified Angular Material theme for better UI consistency. The proxy configuration has also been added to redirect API calls to the backend server. Introduced DTOs, mappers and CRUD operations across the services for better data management and manipulation.