Update styling in articles, handle JWT renewals, edit unique constraint in MddUser model
Updated articles-layout.component.scss and articles-layout.component.html to improve formatting of article titles and subtitles. Adjusted JWT interceptor to handle token expiration by logging out the user. Updated the unique constraint in MddUser.java to not only apply to the email, but also to the username of the user. Additionally, the database password in application.yml has been modified, and two abstract methods have been added to TopicMapper.java.