Implement Swagger and secure controllers
The Swagger tool was implemented to generate documents for the RESTful API and to provide testing features. Security requirements have been added to Post, Comment, Topic, and MddUser controllers and expiry time for tokens has been extended. An endpoint for registering users now returns a message instead of user details. The readme file was updated for better understanding of the project configuration.